backend/requirements.txt
mysql-connector-python==8.0.33
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.0.2
pydantic==2.5.2
pydantic-settings==2.0.3
python-multipart==0.0.6
reportlab==4.0.7
pytz==2023.3
cryptography==41.0.7
